chore: restructure
